Add NTT and the primitive root

+++ algo/math/combinatorics.hpp
@@ -1,7 +1,7 @@
 /*
  * Rożne wzorki kombinatoryczne - digitcrusher/algorytmy
  *
- * Copyright (C) 2021-2022 Karol "digitcrusher" Łacina
+ * Copyright (C) 2021-2023 Karol "digitcrusher" Łacina
  *
  * Copying and distribution of this software, with or without modification,
  * are permitted in any medium without royalty. This software is offered
@@ -15,7 +15,7 @@
  * Liczba k-elementowych podzbiorów n-elementowego
  * zbioru (symbol Newtona) w O(n + log mod)
  */
-ll choose_fac(ll n, ll k, ll mod) {
+ll choose_fac(int n, int k, ll mod) {
   assert(n >= 0 && mod > 0);
   if(k < 0 || k > n) {
     return 0;
@@ -31,7 +31,7 @@ ll choose_fac(ll n, ll k, ll mod) {
  * Liczba k-elementowych podzbiorów n-elementowego zbioru
  * (symbol Newtona) w O(min(k, n - k) log mod)
  */
-ll choose_mul(ll n, ll k, ll mod) {
+ll choose_mul(int n, int k, ll mod) {
   assert(n >= 0 && mod > 0);
   if(k < 0 || k > n) {
     return 0;
@@ -48,14 +48,14 @@ ll choose_mul(ll n, ll k, ll mod) {
   return result;
 }
 
-ll choose(ll n, ll k, ll mod) {
+ll choose(int n, int k, ll mod) {
   return choose_mul(n, k, mod);
 }
 
 /*
  * Liczba sposobów na wsadzenie n nierozróżnialnych elementów do k różnych zbiorów
  */
-ll any_in_uniq(ll n, ll k, ll mod) {
+ll any_in_uniq(int n, int k, ll mod) {
   assert(mod > 0);
   return choose(n + k - 1, k - 1, mod);
 }
@@ -63,7 +63,7 @@ ll any_in_uniq(ll n, ll k, ll mod) {
 /*
  * Liczba sposobów na wsadzenie n różnych elementów do k różnych zbiorów
  */
-ll uniq_in_uniq(ll n, ll k, ll mod) {
+ll uniq_in_uniq(int n, int k, ll mod) {
   assert(mod > 0);
   return mod_mul(any_in_uniq(n, k, mod), fac(n, mod), mod);
 }
